My initial app.py
*My First University Project for Software Design & Architecture

This is a basic initial backend Python app for a banking system. I will continue developing it to make it as good as I can based on my knowledge.

My objective is to create a functional and well-designed banking system using Python as the main backend language, MySQL as the database, and Flask to connect with a WordPress website. In the final version, users will be able to log in to their bank accounts, transfer money, apply for products like credit or deposits, convert currency, and more...